Work in progress on fork.
Store data in sessionStorage partitioned by the state parameter, which uniquely
identifies a SMART app launch instance. This is necessary for certain browser
deployments (like a single-threaded embedded IE instance) in which sessionStorage
is not partitioned naturally by the browser tab.

This work has been submitted to the upstream project but it has not been accepted
yet. See https://github.com/smart-on-fhir/client-js/pull/9 for more information.

All of this work will be contributed back to the open source upstream project.